fix(tests): repair stale imports/renames in 6 test files (WIP toward green dev)
Refactors moved/renamed code without updating tests. Verified each — all are
relocations/renames, coverage preserved (not deletions):
- assistantSkillSelection -> @/renderer/pages/agents/utils
- SkillModalContent fns -> @/renderer/pages/skills/utils
- repairOpenClawConfig -> repairSudoclawConfig (renamed, NOT removed)
- startOpenClawWithRetries -> startSudoclawWithRetries
- mcp messageQueue -> pages/settings/tools/utils/messageQueue (+ mock path)

Part of a larger 27-file / 87-test remediation surfaced by unmasking the
nightly prettier failure. Remaining: DOM render changes, mock-internals
(getHomePage), assertion stale-vs-regression.
